jelly-cinema-v2/jelly-rag-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

langchain-text-splitters>=0.3.0

pymilvus>=2.3.0
redis>=5.0.0
httpx>=0.26.0
python-multipart>=0.0.6
python-docx>=1.1.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1,<4
